requirements.txt
arrow==0.13.1
attrs==18.2.0
backcall==0.1.0
bleach==3.1.0
certifi==2018.11.29
chardet==3.0.4
cssutils==1.0.2
decorator==4.3.2
defusedxml==0.5.0
Django==2.1.5
django-appconf==1.0.3
django-compressor==2.2
django-rest-framework==0.1.0
django-sass-processor==0.7.2
django-webpack-loader==0.6.0
djangorestframework==3.9.2
entrypoints==0.3
idna==2.8
ipdb==0.11
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.2
Jinja2==2.10
jsonschema==3.0.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
libsass==0.18.0
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.4.1
nbformat==4.4.0
notebook==5.7.4
pandocfilters==1.4.2
parso==0.3.2
peewee==3.8.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
prometheus-client==0.6.0
prompt-toolkit==2.0.8
psycopg2==2.7.7
ptyprocess==0.6.0
Pygments==2.3.1
pyrsistent==0.14.11
python-dateutil==2.8.0
pytz==2018.9
pyzmq==18.0.0
qtconsole==4.4.3
rcssmin==1.0.6
requests==2.21.0
rjsmin==1.0.12
Send2Trash==1.5.0
six==1.12.0
terminado==0.8.1
testpath==0.4.2
tinycss2==0.6.1
tornado==5.1.1
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
webcolors==1.8.1
webencodings==0.5.1
widgetsnbextension==3.4.2
